Add `IsFocusable` to the StatefulContentView
